Handling special case of jumps within zero page
Instructions JMP and JSR only have absolute addressing mode.
But with a label within zero page they can get an address with a length of just one byte.
Technically it is now possible to have a 'JMP $A0' instruction which looks like zeropage address mode but is translated to absolute address mode.